+++ src/components/modals/ModalBetResult.js
@@ -1,5 +1,5 @@
 import Modal from 'react-bootstrap/Modal';
-import BetService from "../api/Bet";
+import BetService from "../../api/Bet";
 import { useRef } from 'react';
 
 const ModalBetResult = (props) => {
