+++ src/components/modals/ModalDeleteBet.js
@@ -1,5 +1,5 @@
 import Modal from 'react-bootstrap/Modal';
-import BetService from '../api/Bet';
+import BetService from '../../api/Bet';
 
 const ModalDeleteBet = (props) => {
     function closeBetDeleteModalAndDeleteBet() {
